requirements.txt
nose==1.3.7
Pillow==3.0.0
